requirements.txt
requests
beautifulsoup4
feedparser
openai
pyyaml
pandas